Domain fixes Code reorganization Buisness object "Day" added

+++ app/src/main/java/com/kuznetsova/mlanning/domain/TasksRepository.kt
@@ -6,13 +6,13 @@ import com.kuznetsova.mlanning.domain.taskitem.TaskItem
 interface TasksRepository {
 
     suspend fun getAll(): List<Task>
-    suspend fun getTaskById(id: Int): Task
+    suspend fun getTaskById(id: Long): Task
     suspend fun insertTask(task: Task)
-    suspend fun updateTask(taskId: Int, name: String, description: String?, priority: TaskPriority)
-    suspend fun deleteTask(id: Int)
+    suspend fun updateTask(taskId: Long, name: String, description: String?, priority: TaskPriority)
+    suspend fun deleteTask(id: Long)
 
-    suspend fun updateTaskItem(taskItem: TaskItem, taskId: Int)
-    suspend fun createTaskItem(taskItem: TaskItem, taskId: Int)
+    suspend fun updateTaskItem(taskItem: TaskItem, taskId: Long)
+    suspend fun createTaskItem(taskItem: TaskItem, taskId: Long)
     suspend fun deleteTaskItem(id: Int)
     suspend fun getAllTaskItemByDayId(dayId: Int)
