+++ app/src/main/java/com/kuznetsova/mlanning/domain/task/interactors/DeleteTaskInteractor.kt
@@ -4,7 +4,7 @@ import com.kuznetsova.mlanning.domain.TasksRepository
 
 class DeleteTaskInteractor(private val repository: TasksRepository) {
 
-    suspend fun execute(id: Int) {
+    suspend fun execute(id: Long) {
 
         repository.deleteTask(id)
     }
